requirements.txt
can-isotp==2.0.6
msgpack==1.1.0
packaging==24.2
pyserial==3.5
python-can==4.5.0
typing_extensions==4.12.2
wrapt==1.17.2
